Feature/flatpanel absolute instead of precentage

* Flat Panel Behavior Change - Use absolute flat panel values instead of percentages

* flat device view improvements

* Migrate flat device trained brightness values after first device connection

* Adjusting AUTs for FlatDevices as well as bugfixes revealed by tests

* tick frequency fix

+++ NINA.Core.WPF/Mediator/FlatDeviceMediator.cs
@@ -22,7 +22,7 @@ namespace NINA.WPF.Base.Mediator {
 
     public class FlatDeviceMediator : DeviceMediator<IFlatDeviceVM, IFlatDeviceConsumer, FlatDeviceInfo>, IFlatDeviceMediator {
 
-        public Task SetBrightness(double brightness, CancellationToken token) {
+        public Task SetBrightness(int brightness, CancellationToken token) {
             return handler.SetBrightness(brightness, token);
         }
